Updated search

+++ front_end/app/routes.py
@@ -51,43 +51,38 @@ def pagination(table_data, page):
     return table_data, pages
 
 @app.route('/')
-@app.route('/index/<page>/<colun_name>/<sorting_order>')#main page, show the database and actions you can take
-def index(page = 1, colun_name = 'id', sorting_order = "ASC"):
-    table_data = requests.get(API_LINK, verify=False).json()
+@app.route('/index/<page>')
+@app.route('/index/<page>/<query>')
+@app.route('/index/<page>/<query>/<column_name>')
+@app.route('/index/<page>/<query>/<column_name>/<sorting_order>/')#main page, show the database and actions you can take
+def index(page = 1, query = '*', column_name = 'id', sorting_order = "ASC"):
+    table_data = requests.get(API_LINK + "/GetByAll/" + query , verify=False).json()
     if(table_data == []):
         seed_database()
         table_data = requests.get(API_LINK, verify=False).json()
     table_data, pages = pagination(table_data, int(page))
     return render_template('index.html', title='Home', table_data = table_data, pages = pages, prime_color = PRIME_COLOR, sec_color = SEC_COLOR)
 
-@app.route('/add')#form to add to database
+@app.route("/add", methods=["GET", "POST"])#post the form to the api
 def add():
-    return render_template('add.html', title='Add')
+    responce = None
+    if request.method == 'POST':
+        data = {
+            'name' : request.form['name'],
+            'application' : request.form['application'],
+            'framework' : request.form['framework'],
+            'compatible' : request.form['compatible'],
+        }
+        responce = requests.post(url = API_LINK, json = data,  verify=False).reason
+    return render_template('add.html', title='Add', post_info=responce)
 
-@app.route("/add/post", methods=["GET", "POST"])#post the form to the api
-def add_post():
-    data = {
-        'name' : request.form['name'],
-        'application' : request.form['application'],
-        'framework' : request.form['framework'],
-        'compatible' : request.form['compatible'],
-    }
-    responce = requests.post(url = API_LINK, json = data,  verify=False)
-    return render_template('add.html', title='Post', post_info=responce.reason)
-
-@app.route('/delete')
+@app.route("/delete", methods=["GET", "POST"])#post the delete statement to the api
 def delete():
+    if request.method == 'POST':
+        id = request.form['delete_id']
+        requests.delete(url = API_LINK + "/" + id , verify=False)
     table_data = requests.get(API_LINK, verify=False).json()
     return render_template('delete.html', title='Delete', table_data = table_data, prime_color = PRIME_COLOR, sec_color = SEC_COLOR)
-
-@app.route("/delete/post", methods=["GET", "POST"])#post the delete statement to the api
-def delete_post():
-    id = request.form['delete_id']
-
-    responce = requests.delete(url = API_LINK + "/" + id , verify=False)
-    print(responce)#TODO place this responce somewhere on screen
-    table_data = requests.get(API_LINK, verify=False).json()
-    return render_template('delete.html', title='Post', table_data = table_data, prime_color = PRIME_COLOR, sec_color = SEC_COLOR)
 
 @app.route('/edit')
 def edit():
@@ -101,7 +96,7 @@ def edit_form():
     table_data = requests.get(API_LINK + "/" + id, verify=False).json()
     return render_template('edit_form.html', title='Edit', table_data = table_data)
 
-@app.route('/edit/post', methods=["GET", "POST"])
+@app.route('/edit/post/', methods=["GET", "POST"])#TODO merge with above by usin /<id>
 def edit_form_post():
     data = {
     'id': request.form['id'],
@@ -110,19 +105,16 @@ def edit_form_post():
     'framework' : request.form['framework'],
     'compatible' : request.form['compatible'],
     }
-    responce = requests.put(url = API_LINK + "/" + data["id"], json = data,  verify=False)
-    return render_template('edit_form.html', title='Edit', table_data = data ,post_info=responce.reason)
+    responce = requests.put(url = API_LINK + "/" + data["id"], json = data,  verify=False).reason
+    return render_template('edit_form.html', title='Edit', table_data = data ,post_info=responce)
 
-@app.route('/style')
+@app.route('/style', methods=["GET", "POST"])
 def style():
-    return render_template('style.html', title='Edit', prime_color = PRIME_COLOR, sec_color = SEC_COLOR)
-
-@app.route('/style/post', methods=["GET", "POST"])
-def style_post():
-    global PRIME_COLOR
-    global SEC_COLOR
-    PRIME_COLOR = request.form['prime_color']
-    SEC_COLOR = request.form['sec_color']
+    if request.method == 'POST':
+        global PRIME_COLOR
+        global SEC_COLOR
+        PRIME_COLOR = request.form['prime_color']
+        SEC_COLOR = request.form['sec_color']
     return render_template('style.html', title='Edit', prime_color = PRIME_COLOR, sec_color = SEC_COLOR)
 
 @app.route('/style_rand')
@@ -133,9 +125,14 @@ def style_rand():
     SEC_COLOR = "#{:06x}".format(random.randint(0, 0xFFFFFF))
     return redirect(url_for('index', page=1))
 
-@app.route('/search/<page>')
-def search(page):
-    return render_template('advanced_search.html', title='Search')
+@app.route('/search', methods=["GET", "POST"])
+@app.route('/search/<page>', methods=["GET", "POST"])
+def search(page = 1):
+    query = "*"
+    if request.method == 'POST':
+        query = request.form['search_query']#TODO make sure that this cannot be empty
+    return redirect(url_for('index', query=query, page = page))
+    # return render_template('advanced_search.html', title='Search')
 
 @app.route('/advanced_search')
 def advanced_search():
